Changed power capacity to double

+++ RoboticVacuumCleaner/src/navigation/Sweeper.java
@@ -5,10 +5,10 @@ import util.Debugger;
 public class Sweeper{
 	public static Sweeper sweeper = new Sweeper();
 	private int dirtCapacity;
-	private int powerCapacity;
+	private double powerCapacity;
 	private Sweeper(){
 		dirtCapacity = 50;
-		powerCapacity = 100;
+		powerCapacity = 100.0;
 	}
 	public static Sweeper getInstance(){
 		return sweeper;
@@ -19,11 +19,11 @@ public class Sweeper{
 	public void decreaseDirtCapacity(){
 		dirtCapacity--;
 	}
-	public int checkPowerCapacity(){
+	public double checkPowerCapacity(){
 		return powerCapacity;
 	}
 	
-	public int decreasePowerCapacity(int powerConsumed){
+	public double decreasePowerCapacity(double powerConsumed){
 		return powerCapacity - powerConsumed;
 	}
